refactor: split dynamic handle to swagger and openapi
